requirements.txt
asgiref==3.2.10
awsebcli==3.18.2
botocore==1.15.49
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==2.2.2
django-filter==2.3.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.9.4
djoser==2.0.5
docutils==0.15.2
future==0.16.0
gunicorn==20.0.4
idna==2.7
jmespath==0.10.0
Markdown==3.2.2
pathspec==0.5.9
psycopg2==2.8.6
pypiwin32==223
python-dateutil==2.8.0
python-decouple==3.1
pytz==2020.1
pywin32==228
PyYAML==5.3.1
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
urllib3==1.24.3
wcwidth==0.1.9
whitenoise==5.2.0
